src/data-faker/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.20.*
google-cloud-logging==3.10.*
Faker==18.9.*